Spiking improved resource abstraction - Make NewResourceLoader just an HttpHandler lookup
