+++ http4k-core/src/main/kotlin/org/http4k/routing/experimental/DirectoryResourceLoader.kt
@@ -8,7 +8,7 @@ class DirectoryResourceLoader(
     private val mimeTypes: MimeTypes = MimeTypes()
 ) : NewResourceLoader {
 
-    override fun resourceFor(path: String): FileResource? {
+    override fun invoke(path: String): FileResource? {
         val resolved = baseDir.pathJoin(path.orIndexFile())
         val f = File(resolved)
         return if (!f.exists() || !f.isFile) null else FileResource(f, mimeTypes.forFile(resolved))
