+++ http4k-core/src/test/kotlin/org/http4k/routing/NewStaticRoutingHttpHandlerTest.kt
@@ -1,13 +1,13 @@
 package org.http4k.routing
 
-import org.http4k.routing.experimental.NewResourceLoader
 import org.http4k.routing.experimental.NewStaticRoutingHttpHandler
+import org.http4k.routing.experimental.ResourceLoaders
 
 // I can't move this into experimental for some reason
 
 class NewStaticRoutingHttpHandlerTest : StaticRoutingHttpHandlerTest() {
     override val handler: RoutingHttpHandler = NewStaticRoutingHttpHandler(
         pathSegments = validPath,
-        resourceLoader = NewResourceLoader.Classpath()
+        resourceLoader = ResourceLoaders.Classpath()
     )
 }
